changes : int64 data model to int

+++ entity/submission.go
@@ -6,10 +6,10 @@ import (
 )
 
 type Submission struct {
-	ID           int64           `json:"id" gorm:"primaryKey"`
-	UserId       int64           `json:"user_id"`
+	ID           int             `json:"id" gorm:"primaryKey"`
+	UserId       int             `json:"user_id"`
 	Answer       json.RawMessage `json:"answer" gorm:"type:jsonb"`
-	RiskScore    int64           `json:"risk_score"`
+	RiskScore    int             `json:"risk_score"`
 	RiskCategory string          `json:"risk_category"`
 	CreatedAt    time.Time       `json:"created_at"`
 	UpdatedAt    time.Time       `json:"updated_at"`
